fix(linux): Fix linux app detection.
Installed state detection fix.
Linux download location fixes
