+++ app/actions/alias/app_manager_actions.ts
@@ -143,6 +143,7 @@ const getLatestAppVersions = async (): Promise<void> => {
             );
 
             if ( isAppInstalledLocally ) {
+                updatedApp.isInstalled = true;
                 ipcRenderer.send( 'checkApplicationsForUpdate', updatedApp );
             }
             store.dispatch( updateAppInfoIfNewer( updatedApp ) );
